fix: 답변 API 질문 식별자를 turn 기준으로 복구
답변 제출과 평가 응답에서 questionId를 Message.id가 아니라 문항 turn 값으로 통일해 API 계약 혼선을 제거했습니다.

Constraint: 최용성님 확인에 따라 요청/응답 DTO 모두 turn 기준을 사용함

Rejected: Message.id와 turn을 서비스에서 상호 변환하는 방식 | API 계약이 실제 질문 식별자와 turn을 혼용하게 됨

Confidence: high

Scope-risk: moderate

Directive: interview API의 questionId는 문항 순서 turn 값으로 해석해야 함

Tested: .\\gradlew.bat test

Not-tested: 없음

+++ src/main/java/com/careerdungeon/domain/interview/dto/InterviewAnswerEvaluationResponse.java
@@ -1,7 +1,7 @@
 package com.careerdungeon.domain.interview.dto;
 
 public record InterviewAnswerEvaluationResponse(
-        Long questionId,
+        Integer questionId,
         int score,
         String feedback
 ) {
