+++ src/main/java/com/careerdungeon/domain/interview/dto/InterviewQuestionResponse.java
@@ -1,7 +1,7 @@
 package com.careerdungeon.domain.interview.dto;
 
 public record InterviewQuestionResponse(
-        Long questionId,
+        Integer questionId,
         String question
 ) {
 }
